+++ src/c/ncm_conversion.c
@@ -175,7 +175,7 @@ ncm_conversion_set_f64_lower_error(NcmError *error, double value,
 
 bool
 ncm_parse_int64(char *source, int32 source_len,
-                int64 *out, NcmError *error) {
+                int32 *out, NcmError *error) {
     NcmBuffer buffer;
     char *end;
     uint64 value;
@@ -200,7 +200,7 @@ ncm_parse_int64(char *source, int32 source_len,
          && (errno != ERANGE)
          && (value <= (uint64)MAXOF(*out));
     if (ok) {
-        *out = (int64)value;
+        *out = (int32)value;
         ncm_error_clear(error);
     } else {
         ncm_conversion_set_parse_error(error, source, source_len);
@@ -212,7 +212,7 @@ ncm_parse_int64(char *source, int32 source_len,
 
 bool
 ncm_parse_int32(char *source, int32 source_len, int32 *out, NcmError *error) {
-    int64 value;
+    int32 value;
 
     if (out == NULL) {
         ncm_error_set(error, EINVAL, STRLIT_ARGS("missing conversion output"));
